Enable test-integration-cli for Windows platform
