+++ integration-cli/docker_cli_daemon_test.go
@@ -1,3 +1,5 @@
+// +build daemon
+
 package main
 
 import (
