+++ integration-cli/docker_cli_run_test.go
@@ -20,9 +20,7 @@ import (
 	"time"
 
 	"github.com/docker/docker/nat"
-	"github.com/docker/docker/pkg/mount"
 	"github.com/docker/docker/pkg/networkfs/resolvconf"
-	"github.com/kr/pty"
 )
 
 // "test123" should be printed by docker run
@@ -1240,45 +1238,6 @@ func TestRunDisallowBindMountingRootToRoot(t *testing.T) {
 	logDone("run - bind mount /:/ as volume should fail")
 }
 
-// Test recursive bind mount works by default
-func TestRunWithVolumesIsRecursive(t *testing.T) {
-	tmpDir, err := ioutil.TempDir("", "docker_recursive_mount_test")
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	defer os.RemoveAll(tmpDir)
-
-	// Create a temporary tmpfs mount.
-	tmpfsDir := filepath.Join(tmpDir, "tmpfs")
-	if err := os.MkdirAll(tmpfsDir, 0777); err != nil {
-		t.Fatalf("failed to mkdir at %s - %s", tmpfsDir, err)
-	}
-	if err := mount.Mount("tmpfs", tmpfsDir, "tmpfs", ""); err != nil {
-		t.Fatalf("failed to create a tmpfs mount at %s - %s", tmpfsDir, err)
-	}
-	defer mount.Unmount(tmpfsDir)
-
-	f, err := ioutil.TempFile(tmpfsDir, "touch-me")
-	if err != nil {
-		t.Fatal(err)
-	}
-	defer f.Close()
-
-	runCmd := exec.Command(dockerBinary, "run", "--name", "test-data", "--volume", fmt.Sprintf("%s:/tmp:ro", tmpDir), "busybox:latest", "ls", "/tmp/tmpfs")
-	out, stderr, exitCode, err := runCommandWithStdoutStderr(runCmd)
-	if err != nil && exitCode != 0 {
-		t.Fatal(out, stderr, err)
-	}
-	if !strings.Contains(out, filepath.Base(f.Name())) {
-		t.Fatal("Recursive bind mount test failed. Expected file not found")
-	}
-
-	deleteAllContainers()
-
-	logDone("run - volumes are bind mounted recursively")
-}
-
 func TestRunDnsDefaultOptions(t *testing.T) {
 	// ci server has default resolv.conf
 	// so rewrite it for the test
@@ -2281,44 +2240,6 @@ func TestRunExecDir(t *testing.T) {
 	}
 
 	logDone("run - check execdriver dir behavior")
-}
-
-// #6509
-func TestRunRedirectStdout(t *testing.T) {
-
-	defer deleteAllContainers()
-
-	checkRedirect := func(command string) {
-		_, tty, err := pty.Open()
-		if err != nil {
-			t.Fatalf("Could not open pty: %v", err)
-		}
-		cmd := exec.Command("sh", "-c", command)
-		cmd.Stdin = tty
-		cmd.Stdout = tty
-		cmd.Stderr = tty
-		ch := make(chan struct{})
-		if err := cmd.Start(); err != nil {
-			t.Fatalf("start err: %v", err)
-		}
-		go func() {
-			if err := cmd.Wait(); err != nil {
-				t.Fatalf("wait err=%v", err)
-			}
-			close(ch)
-		}()
-
-		select {
-		case <-time.After(10 * time.Second):
-			t.Fatal("command timeout")
-		case <-ch:
-		}
-	}
-
-	checkRedirect(dockerBinary + " run -i busybox cat /etc/passwd | grep -q root")
-	checkRedirect(dockerBinary + " run busybox cat /etc/passwd | grep -q root")
-
-	logDone("run - redirect stdout")
 }
 
 // Regression test for https://github.com/docker/docker/issues/8259
