+++ integration-cli/docker_test_vars.go
@@ -4,6 +4,7 @@ import (
 	"fmt"
 	"os"
 	"os/exec"
+	"runtime"
 )
 
 var (
@@ -25,16 +26,26 @@ var (
 	workingDirectory string
 )
 
+func binarySearchCommand() *exec.Cmd {
+	if runtime.GOOS == "windows" {
+		// Windows where.exe is included since Windows Server 2003. It accepts
+		// wildcards, which we use here to match the development builds binary
+		// names (such as docker-$VERSION.exe).
+		return exec.Command("where.exe", "docker*.exe")
+	}
+	return exec.Command("which", "docker")
+}
+
 func init() {
 	if dockerBin := os.Getenv("DOCKER_BINARY"); dockerBin != "" {
 		dockerBinary = dockerBin
 	} else {
-		whichCmd := exec.Command("which", "docker")
+		whichCmd := binarySearchCommand()
 		out, _, err := runCommandWithOutput(whichCmd)
 		if err == nil {
 			dockerBinary = stripTrailingCharacters(out)
 		} else {
-			fmt.Printf("ERROR: couldn't resolve full path to the Docker binary")
+			fmt.Printf("ERROR: couldn't resolve full path to the Docker binary (%v)", err)
 			os.Exit(1)
 		}
 	}
